feat(submodules/html2pdf): update to the latest version
WHAT:

The upstream JS library has fixed the following issue:

If a paragraph contains only a single bold text fragment and the rest of the
text is not wrapped, the strong:only-child selector still matches. This is a
known limitation and is currently not solved at the html2pdf level.

WHY:

This fixes a user bug report:
https://github.com/strictdoc-project/strictdoc/discussions/2593#discussioncomment-15253312

HOW:

- https://github.com/mettta/html2pdf4doc/pull/174
- https://github.com/mettta/html2pdf4doc/pull/177